fix: precision loss for xprotocol clientStreamIDBase (#2404)
* fix: precision loss for xprotocol clientStreamIDBase

* fix: file header of license

(cherry picked from commit febc9a879b5c1499c2d11eea569f535d8901a5ef)

+++ pkg/protocol/xprotocol/boltv2/protocol.go
@@ -213,5 +213,6 @@ func (proto boltv2Protocol) EnableWorkerPool() bool {
 }
 
 func (proto boltv2Protocol) GenerateRequestID(streamID *uint64) uint64 {
-	return atomic.AddUint64(streamID, 1)
+	// fix bug for issue: https://github.com/mosn/mosn/issues/2403
+	return uint64(uint32(atomic.AddUint64(streamID, 1)))
 }
